Added notifications feature

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -59,7 +59,19 @@ export const POST = async (
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.upsert({
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    })
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.upsert({
         where: {
             userId_postId: {
                 userId: loggedInUser.id,
@@ -71,7 +83,19 @@ export const POST = async (
             postId
         },
         update: {}
-    })
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [prisma.notification.create({
+              data: {
+              issuerId: loggedInUser.id,
+              recipientId: post.userId,
+              postId,
+              type: "LIKE",
+            },
+          })]
+        : []
+      )
+    ]);
 
     return new Response();
   } catch (error) {
@@ -91,12 +115,33 @@ export const DELETE = async (
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    const post = await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
-    });
+    })
+
+    if (!post) {
+      return Response.json({ error: "Post not found" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      })
+    ])
 
     return new Response();
   } catch (error) {
